full funtion and add admin login

+++ src/main/java/com/chrima_chris_game/controller/AdminController.java
@@ -8,6 +8,8 @@ import com.chrima_chris_game.entity.Player;
 import com.chrima_chris_game.service.AssignmentService;
 import com.chrima_chris_game.service.PlayerService;
 
+import jakarta.servlet.http.HttpSession;
+
 
 @RestController
 @RequestMapping("/admin")
@@ -24,33 +26,45 @@ public class AdminController {
 
 
     @PostMapping("/addPlayer")
-    public String addPlayer(@RequestBody Player player) {
+    public String addPlayer(@RequestBody Player player,HttpSession session) {
+    	checkAdmin(session);
         return playerService.addPlayer(player);
     }
     
     @PostMapping("/startGame")
-    public String startGame() {
+    public String startGame(HttpSession session) {
+    	checkAdmin(session);
         return assignmentService.startGame();
     }
     
     @PostMapping("/resetGame")
-    public String resetGame() {
+    public String resetGame(HttpSession session) {
+    	checkAdmin(session);
         return assignmentService.resetGame();
     }
     
     @PostMapping("/fullResetGame")
-    public String fullResetGame() {
+    public String fullResetGame(HttpSession session) {
+    	checkAdmin(session);
         return assignmentService.fullResetGame();
     }
     
     @GetMapping("/players")
-    public List<Player> getAllPlayers() {
+    public List<Player> getAllPlayers(HttpSession session) {
+    	checkAdmin(session);
         return playerService.getAllPlayers();
     }
     
     @DeleteMapping("/deletePlayer/{id}")
-    public String deletePlayer(@PathVariable int id) {
+    public String deletePlayer(@PathVariable int id,HttpSession session) {
+    	checkAdmin(session);
         return playerService.deletePlayer(id);
+    }
+
+    private void checkAdmin(HttpSession session) {
+        if (session.getAttribute("ADMIN") == null) {
+            throw new RuntimeException("Unauthorized");
+        }
     }
